Merging db layout with main

+++ frontend/src/components/admin/DeleteSeason.js
@@ -5,16 +5,27 @@ const DeleteSeason = ({ seasonName, onComplete }) => {
     const handleDelete = async () => {
         const confirmDelete = window.confirm(`Are you sure you want to delete the season: ${seasonName}?`);
         if (confirmDelete) {
+            const token = localStorage.getItem('access_token');  // Get token from localStorage
+            if (!token) {
+                alert('Unauthorized: No token found');
+                return;
+            }
+
             try {
                 const response = await fetch(`http://localhost:8000/clear_season/`, {
                     method: 'DELETE',
                     headers: {
-                        'Content-Type': 'application/json'
+                        'Authorization': `Bearer ${token}`,  // Pass token in Authorization header
+                        'Content-Type': 'application/json',
                     },
                 });
+
                 if (response.ok) {
                     alert(`Season '${seasonName}' deleted successfully`);
-                    onComplete(); // Exit delete mode
+                    // Check if onComplete is a function before calling it
+                    if (typeof onComplete === 'function') {
+                        onComplete(); // Exit delete mode
+                    }
                 } else {
                     alert('Failed to delete season');
                 }
